Create wpvip-polylang-term-count.php
In a WP VIP hosting environment, term counting is deferred. This breaks Polylang's associations between terms and posts, so that translations cannot be related to each other if the site has more than two languages.

This solution forces the term count to flush when a new post is associated with a translation term.